Continue debugging, problem identified
The host container spawns a sibling container and tries to mount a
directory from the host container filesystem onto the sibling container
filesystem. However, the docker daemon is running on the host machine,
not the host container. Thus, all mount paths are on the host machine
filesystem. The solution can be:
- create /var/lib/featurefactory/data on the host machine
- mount /var/lib/featurefactory:/var/lib/featurefactory on the host container